+++ views/admin/message/index.php
@@ -9,8 +9,9 @@ require_once ROOT . '/views/include/header-admin.php';
 		<h1 class="h3 mb-0 text-gray-800">Сообщения</h1>
 		<a href="#" class="d-none d-sm-inline-block btn btn-sm btn-primary shadow-sm"><i class="fas fa-download fa-sm text-white-50"></i> Generate Report</a>
 	</div>
-	<p class="mb-4">DataTables is a third party plugin that is used to generate the demo table below. For more information about DataTables, please visit the <a target="_blank" href="https://datatables.net">official DataTables documentation</a>.</p>
-	<div class="card shadow mb-4">
+	<p class="mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Adipisci commodi consequatur dolores dolorum est exercitationem hic in inventore ipsa numquam obcaecati praesentium provident quidem quos repellat repellendus, repudiandae, similique totam?</p>
+
+    <div class="card shadow mb-4">
 		<div class="card-header py-3">
 			<h6 class="m-0 font-weight-bold text-primary">Сообшение</h6>
 		</div>
@@ -47,7 +48,7 @@ require_once ROOT . '/views/include/header-admin.php';
 						<td><?=$row['dateadd']?></td>
 						<td>
                             <a href="/admin/message/view/<?=$row['id']?>" class="btn btn-primary"><i class="fas fa-eye"></i></a>
-                            <a href="/views/delete.php?id=<?=$row['id']?>&type=message" class="btn btn-danger"><i class="fas fa-trash-alt"></i></a>
+                            <a href="/views/delete.php?id=<?=$row['id']?>&type=message" onclick="return confirm('Вы действительно хотите удалить?')" class="btn btn-danger"><i class="fas fa-trash-alt"></i></a>
                         </td>
 					</tr>
                     <?php endforeach; ?>
